collector/requirements.txt
#
# requirements.txt for git-branch-metrics-collector.py
#
# sh
gitpython
